ingest/requirements.txt
google-cloud-bigquery
google-cloud-storage
pytest